+++ packages/xrpl/src/utils/index.ts
@@ -20,6 +20,7 @@ import {
   encodeForMultisigning as rbcEncodeForMultisigning,
   encodeForSigning as rbcEncodeForSigning,
   encodeForSigningClaim as rbcEncodeForSigningClaim,
+  encodeForSigningBatch as rbcEncodeForSigningBatch,
 } from 'ripple-binary-codec'
 import { verify as verifyKeypairSignature } from 'ripple-keypairs'
 
@@ -46,6 +47,9 @@ import {
   hashLedgerHeader,
   hashEscrow,
   hashPaymentChannel,
+  hashVault,
+  hashLoanBroker,
+  hashLoan,
 } from './hashes'
 import parseNFTokenID from './parseNFTokenID'
 import {
@@ -125,6 +129,16 @@ function encodeForMultiSigning(object: Transaction, signer: string): string {
 }
 
 /**
+ * Encodes a Batched Transaction for multi-account signing
+ *
+ * @param object - Batched Transaction in JSON format.
+ * @returns A hex string representing the encoded object.
+ */
+function encodeForSigningBatch(object: Transaction): string {
+  return rbcEncodeForSigningBatch(object)
+}
+
+/**
  * Decodes a hex string into a transaction | ledger entry
  *
  * @param hex - hex string in the XRPL serialization format.
@@ -179,6 +193,9 @@ const hashes = {
   hashLedgerHeader,
   hashEscrow,
   hashPaymentChannel,
+  hashVault,
+  hashLoanBroker,
+  hashLoan,
 }
 
 export {
@@ -225,6 +242,7 @@ export {
   decode,
   encodeForMultiSigning,
   encodeForSigning,
+  encodeForSigningBatch,
   encodeForSigningClaim,
   getNFTokenID,
   parseNFTokenID,