+++ packages/xrpl/HISTORY.md
@@ -4,6 +4,12 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ## Unreleased
 
+## 4.5.0 (2025-12-16)
+
+### Added
+* Support for `Lending Protocol` (XLS-66d).
+* Export signing and binary codec utilities.
+
 ### Fixed
 * Update ripple-binary-codec to 2.5.1 to address serialization/deserialization issues in `Issue` serialized type for `MPTIssue`.
 * Better faucet error handling